Loosen types for a more generic experience
